requirements.txt
selenium==4.2.0
allure-pytest==2.9.45
selene==2.0.0b5
webdriver-manager==3.8.5
pytest~=6.2.5
pyyaml==5.4.1
requests~=2.26.0
python-dotenv==0.19.0
Cerberus~=1.3.4
PyJWT~=2.3.0
pylint==2.13.9
pylint-gitlab==1.0.0
discord==1.7.3
discord.py==1.7.3
pypdf2==2.11.2
urllib3~=1.26.12